Use pattern-matchers for generating JSON too.
I'm not too happy with this yet.  There are quite a few logic errors
that should be statically detectable yet aren't.  Unfortunately my first
attempt at trying to fix them ran into the limitations of the type
inferencer.

Still, at least it does the right thing when the code is correct.