JSON only has "number"; remove JIntegral and JFloatingPoint
The distinction doesn't sit easily with case classes.  Is "5" an integer
or a float that happens to have had an integer value and no decimal
point printed?  And JIntegral can't comfortably be a subclass of
JFloatingPoint (even if case class inheritance weren't deprecated) because
there are Long values that can't be represented in Doubles.  So just
punt it to a higher level -- JNumber wraps a BigDecimal and it's user
code's responsibility for extracting a primitive if that's what's wanted.