New BlockTokenIterator and less copying in the JsonLexer
So it turns out that the immutable JsonLexer is actually significantly
faster than the Reader-based lexer.  Go figure.  So here's a lexer-
iterator that can be used when you know you're going to be consuming
the entire contents of a Reader.

Also reduce copying even more by the introduction of the WrappedCharArray
which is like a lightweight read-only CharBuffer.  It also uses
reflection tricks to wrap the contents of Strings without copying them,
so there's that too.